Petshop/rest_petshop/serializers.py
from rest_framework import serializers
from core.models import Producto

class ProductoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Producto
        fields = ['nombre','marca','descripcion','categoria','stock','precio','imagenProducto']